JAMES-2540 Split RabbitMQ Mail Queue class in smaller chuncks
Have a package-visible class for each mail queue operation to help readability
